添加更新版本说明

+++ src/main/java/demo/tcyeee/controller/planTask/PlanTaskController.java
@@ -2,13 +2,12 @@ package demo.tcyeee.controller.planTask;
 
 import demo.tcyeee.entity.po.PlanTask;
 import demo.tcyeee.service.PlanTaskService;
+import io.micrometer.core.instrument.util.StringUtils;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
 import javax.annotation.Resource;
-import javax.validation.constraints.NotEmpty;
 
-import static demo.tcyeee.entity.enums.base.ReturnCodeList.*;
 import static demo.tcyeee.entity.enums.base.ReturnCodeList.ReturnCode.*;
 import static demo.tcyeee.utils.ResponseUtils.*;
 
@@ -33,7 +32,11 @@ public class PlanTaskController {
      * @return status
      */
     @RequestMapping("creatTask")
-    public String creatTask(@NotEmpty(message = PARAMS_ERROR_INFO + "task") PlanTask task) {
+    public String creatTask(PlanTask task) {
+        if (task == null || StringUtils.isBlank(task.getName())) {
+            return creatErrResponse(PARAMS_ERROR);
+        }
+
         boolean creatTask = planTaskService.creatTask(task);
         return creatTask ? creatJsonResponse(SUCCESS) : creatErrResponse(SYSTEM_ERROR);
     }
@@ -57,8 +60,10 @@ public class PlanTaskController {
      * @return status
      */
     @RequestMapping("updateTask")
-    public String updateTask(@NotEmpty(message = PARAMS_ERROR_INFO + "task") PlanTask task) {
-        if (task.getId() == null) return creatErrResponse(PARAMS_ERROR);
+    public String updateTask(PlanTask task) {
+        if (task == null || task.getId() == null) {
+            return creatErrResponse(PARAMS_ERROR);
+        }
 
         boolean update = planTaskService.update(task);
         return update ? creatJsonResponse(SUCCESS) : creatErrResponse(SYSTEM_ERROR);
@@ -71,7 +76,9 @@ public class PlanTaskController {
      * @return status
      */
     @RequestMapping("deleteTask")
-    public String deleteTask(@NotEmpty(message = PARAMS_ERROR_INFO + "taskId") Integer taskId) {
+    public String deleteTask(Integer taskId) {
+        if (taskId == null) return creatErrResponse(PARAMS_ERROR);
+
         planTaskService.delete(taskId);
         return creatJsonResponse(SUCCESS);
     }
